requirements.txt
python-telegram-bot
ccxt
pandas
pandas_ta
aiosqlite
python-dotenv
numpy
APScheduler